+++ HISTORY.md
@@ -60,6 +60,7 @@
 ### 2022-07-18
 * German translation update: Changed translation of "point in direction" to "setze Richtung auf ... Grad"
 * threads: JSCompiler scope fixes and redesign, thanks, xBZZZZ!
+* blocks: added support for mixed-typed polyadic inputs 
 
 ### 2022-07-11
 * extensions: fixed #3065